requirements.txt
amqp==2.2.2
anyjson==0.3.3
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.1.0
billiard==3.5.0.3
Brlapi==0.6.6
cached-property==1.5.1
celery==4.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
cupshelpers==1.0
defer==1.0.6
defusedxml==0.6.0
distro-info==0.18
Django==2.2.2
django-computed-property==0.2.4
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
ephem==3.7.6.0
httplib2==0.9.2
idna==2.8
isodate==0.6.0
itypes==1.1.0
Jinja2==2.10.1
jwt==0.6.1
keyring==10.6.0
keyrings.alt==3.0
Khayyam==3.0.17
kombu==4.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.3.4
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
python-dateutil==2.6.1
python-debian==0.1.32
python-memcached==1.53
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.3
usb-creator==0.3.3
vine==1.1.4
virtualenv==16.6.0
wadllib==1.3.2
xkit==0.0.0
zeep==3.4.0
zope.interface==4.3.2
